# agent/src/baseliner_agent/tests/openapi.pretty.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Baseliner API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/admin/enroll-tokens": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Create Enroll Token",
                "operationId": "create_enroll_token_api_v1_admin_enroll_tokens_post",
                "parameters": [
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEnrollTokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateEnrollTokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/assign-policy": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Assign Policy",
                "operationId": "assign_policy_api_v1_admin_assign_policy_post",
                "parameters": [
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignPolicyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssignPolicyResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/policies": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Upsert Policy",
                "operationId": "upsert_policy_api_v1_admin_policies_post",
                "parameters": [
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpsertPolicyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpsertPolicyResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/devices": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "List Devices",
                "operationId": "list_devices_api_v1_admin_devices_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 50,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "include_health",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "description": "If true, include last_run + computed health fields per device.",
                            "default": false,
                            "title": "Include Health"
                        },
                        "description": "If true, include last_run + computed health fields per device."
                    },
                    {
                        "name": "stale_after_seconds",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 86400,
                            "minimum": 60,
                            "description": "Mark device as 'stale' if latest run is older than this many seconds.",
                            "default": 1800,
                            "title": "Stale After Seconds"
                        },
                        "description": "Mark device as 'stale' if latest run is older than this many seconds."
                    },
                    {
                        "name": "offline_after_seconds",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 604800,
                            "minimum": 60,
                            "description": "Mark device as 'offline' if last_seen_at is older than this many seconds.",
                            "default": 3600,
                            "title": "Offline After Seconds"
                        },
                        "description": "Mark device as 'offline' if last_seen_at is older than this many seconds."
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DevicesListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/runs": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "List Runs",
                "operationId": "list_runs_api_v1_admin_runs_get",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Device Id"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 50,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RunsListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/runs/{run_id}": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Get Run Detail",
                "operationId": "get_run_detail_api_v1_admin_runs__run_id__get",
                "parameters": [
                    {
                        "name": "run_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Run Id"
                        }
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RunDetailResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AssignPolicyRequest": {
                "properties": {
                    "device_id": {
                        "type": "string",
                        "title": "Device Id"
                    },
                    "policy_name": {
                        "type": "string",
                        "title": "Policy Name"
                    },
                    "mode": {
                        "type": "string",
                        "title": "Mode",
                        "default": "enforce"
                    },
                    "priority": {
                        "type": "integer",
                        "title": "Priority",
                        "default": 100
                    }
                },
                "type": "object",
                "required": [
                    "device_id",
                    "policy_name"
                ],
                "title": "AssignPolicyRequest"
            },
            "AssignPolicyResponse": {
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "title": "Ok"
                    }
                },
                "type": "object",
                "required": [
                    "ok"
                ],
                "title": "AssignPolicyResponse"
            },
            "CreateEnrollTokenRequest": {
                "properties": {
                    "expires_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Expires At"
                    },
                    "note": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Note"
                    }
                },
                "type": "object",
                "title": "CreateEnrollTokenRequest"
            },
            "CreateEnrollTokenResponse": {
                "properties": {
                    "enroll_token": {
                        "type": "string",
                        "title": "Enroll Token"
                    },
                    "expires_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Expires At"
                    }
                },
                "type": "object",
                "required": [
                    "enroll_token"
                ],
                "title": "CreateEnrollTokenResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "UpsertPolicyRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 3,
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "schema_version": {
                        "type": "string",
                        "title": "Schema Version",
                        "default": "1.0"
                    },
                    "document": {
                        "type": "object",
                        "title": "Document"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "document"
                ],
                "title": "UpsertPolicyRequest"
            },
...
        }
    }
}